ci: move linters into their own workflow
Move the remaining rust-ci workflow to a linter workflow and split the
linters into separate jobs to allow them to run in parallel and provide
output of both, even if one fails.

Similarly to the unittests, allow the workflows to run on `wip/*`
branches for private forks. Lastly, add some more doc-strings to make
the output easier to recognize in the UI.

Signed-off-by: David Rheinsberg <david.rheinsberg@gmail.com>